[Reland] Sparse Conditional Constant Propagation
- This generates better code on some benchmark that have branches
controlled by constants. Nevertheless on real application is does
not provide real benefits, it saves only few bytes.
- This algorithm seems to slighty impact compilation time by
around 1,6%, it represent less than 500ms in GMSCore v10.

Bug: 69293928

+++ src/main/java/com/android/tools/r8/ir/code/If.java
@@ -162,8 +162,7 @@ public class If extends JumpInstruction {
     return 0;
   }
 
-
-  public BasicBlock targetFromCondition(int cond) {
+  public BasicBlock targetFromCondition(long cond) {
     switch (type) {
       case EQ:
         return cond == 0 ? getTrueTarget() : fallthroughBlock();
